-made some fixes in cart controller -made some changes in user schema and cart schema

+++ src/controllers/cartController.js
@@ -1,39 +1,35 @@
 const Cart = require("../models/cart");
+const Food = require("../models/food");
 const User = require("../models/user");
 const yup = require("yup");
 
 const addCart = async (req, res) => {
  try{
-   const userId = req?.query?.userId;
+   const userId = req?.userId;
    const foodId = req?.query?.foodId;
-   
-  const { no_of_item ,cartStaus } = req?.body;
 
   if(!userId || !foodId){
     return res.status(400).send({ message: "user id or food Id not found" });
   }
 
-  const addCartSchema = yup.object({
-    no_of_item: yup.number().required('no. of items are required'),
-    cartStaus: yup.string().required('cart status is required'),
-   })
+  const findFood = await Food.findOne({ where: { id: foodId } });
 
-  await addCartSchema.validate({ no_of_item, cartStaus });
+  if(!findFood){
+    return res.status(404).send({ message: 'no food exists' });
+  }
 
   const response = await User.findOne({
     where:{
       id: userId,
-      attribute:[cart_code]
-    }
-  })
+    },
+    attribute:['cart_code']
+  });
 
   const cartData = {
     userId,
     foodId,
-    no_of_item,
-    cartStaus,
-    cart_code:response,
-  }
+    cart_code:response?.cart_code,
+  };
 
   if(response){
     const response2 = await Cart.create(cartData);
@@ -57,16 +53,16 @@ const updateCart = async (req, res) => {
     const { ...data } = req?.body;
 
     if(!cartId){
-      return res.status(400).send({ message: "Cart Id not found"})
+      return res.status(400).send({ message: "Cart Id not found"});
     }
 
     const updateCartSchema = yup.object({
       no_of_item: yup.number().required('No of itesm is requires'),
     })
+    
+    await updateCartSchema.validate(data);
 
-    await updateCartSchema.validate({ data })
-
-    const response = await Cart.update(data, {id: cartId})
+    const response = await Cart.update(data, {where: { id: cartId }})
 
     if(response){
       return res.status(200).send({message: "cart updated successfully!!"});
@@ -74,31 +70,31 @@ const updateCart = async (req, res) => {
       return res.status(400).send({ message: "Something occured while updating cart" });
     }
   }catch(err){
+    console.log(err)
     return res.status(500).status({ message: err.message || "Internal Server Error" }) ;
   }
 }
 
 const getCart = async (req, res) => {
   try{
-    const userId = req?.query?.userId;
+    const userId = req?.userId;
 
     if(!userId){
       return res.status(400).send({ message: "user id not found" })
     }
 
-    const response = await User.findOne({ where: { id: userId , attribute: [cart_code] } });
+    const response = await User.findOne({ where: { id: userId }, attribute: [`cart_code`] });
 
     if(response){
-      const response2 = await Cart.findAll({ where: { cart_code: response.cart_code } });
-
-      if(response2){
+      const response2 = await Cart.findAll({ where: { cart_code: response?.cart_code } , include: [Food] });
+      
+      if(response2?.length){
         return res.status(200).send({ message: "Successfully fetched data", data: response2 });
-      }else{
-        return res.status(400).status({ message: 'No Cart Data Found' });
       }
     }else{
       return res.status(400).send({ message: "No User Data Found" });
     }
+    
   }catch(err){
     return res.status(500).send({ message: err.message || "Internal Server Error" });
   }
@@ -107,15 +103,17 @@ const getCart = async (req, res) => {
 const deleteCartItem = async (req, res) => {
   try{
     const foodId = req?.query?.foodId;
+    const cartId = req?.query?.cartId;
+    const userId = req?.userId; 
 
-    if(!foodId){
-      return res.status(400).send({ message: "cart id not found" });
+    if(!foodId || !cartId || !userId){
+      return res.status(400).send({ message: "cart id or food id or user id not found" });
     }
 
-    const response = await Cart.destroy({ where: { foodId } });
+    const response = await Cart.destroy({ where: { foodId, userId, id: cartId } });
 
     if(response){
-      return res.status(200).send({ message: "Successfully delted cart item" });
+      return res.status(200).send({ message: "Successfully item removed" });
     }else{
       return res.status(400).send({ message: "Something Occured while deleting cart" });
     }
